backend/requirements.txt
Django==5.2.3
djangorestframework==3.16.1
django-cors-headers
djangorestframework-simplejwt==5.5.1
mssql-django==1.6
pyodbc==5.3.0
drf-spectacular
